RandomSystem Class 생성 및 적용

+++ Assets/Script/02_Objects/00_Common/ObjectConstant.cs
@@ -1,5 +1,5 @@
 public class ObjectConstant
 {
     public const int SoulIdleMotionSpeed = 100; // 느낌으로 괜찮은 값을 찾은 것 - 2024.11.21 신동환
-    public const int SoulDropRate = 20; // %
+    public const float SoulDropRate = 20.0f; // %
 }
